Added save/restore to client-side VScript
