+++ sp/src/game/client/c_world.h
@@ -31,6 +31,7 @@ public:
 
 	virtual void Precache();
 	virtual void Spawn();
+	virtual bool KeyValue( const char *szKeyName, const char *szValue );
 
 	// Don't worry about adding the world to the collision list; it's already there
 	virtual CollideType_t	GetCollideType( void )	{ return ENTITY_SHOULD_NOT_COLLIDE; }
@@ -41,8 +42,15 @@ public:
 	float GetWaveHeight() const;
 	const char *GetDetailSpriteMaterial() const;
 
+#ifdef MAPBASE
+	// A special function which parses map data for the client world entity before LevelInitPreEntity().
+	// This can be used to access keyvalues early and without transmitting from the server.
+	void ParseWorldMapData( const char *pMapData );
+#endif
+
 #ifdef MAPBASE_VSCRIPT
-	ScriptLanguage_t GetScriptLanguage() { return (ScriptLanguage_t)m_iScriptLanguageClient; }
+	// -2 = Use server language
+	ScriptLanguage_t GetScriptLanguage() { return (ScriptLanguage_t)(m_iScriptLanguageClient != -2 ? m_iScriptLanguageClient : m_iScriptLanguageServer); }
 #endif
 
 public:
@@ -64,6 +72,7 @@ public:
 	char	m_iszChapterTitle[64];
 #endif
 #ifdef MAPBASE_VSCRIPT
+	int		m_iScriptLanguageServer;
 	int		m_iScriptLanguageClient;
 #endif
