+++ sp/src/game/server/vscript_server.h
@@ -15,19 +15,7 @@
 #pragma once
 #endif
 
-class ISaveRestoreBlockHandler;
-
 bool VScriptServerReplaceClosures( const char *pszScriptName, HSCRIPT hScope, bool bWarnMissing = false );
-ISaveRestoreBlockHandler *GetVScriptSaveRestoreBlockHandler();
-
-
-class CBaseEntityScriptInstanceHelper : public IScriptInstanceHelper
-{
-	bool ToString( void *p, char *pBuf, int bufSize );
-	void *BindOnRead( HSCRIPT hInstance, void *pOld, const char *pszId );
-};
-
-extern CBaseEntityScriptInstanceHelper g_BaseEntityScriptInstanceHelper;
 
 // Only allow scripts to create entities during map initialization
 bool IsEntityCreationAllowedInScripts( void );
