+++ sp/src/game/server/world.cpp
@@ -395,7 +395,7 @@ BEGIN_DATADESC( CWorld )
 	DEFINE_KEYFIELD( m_iszDetailSpriteMaterial, FIELD_STRING, "detailmaterial" ),
 #ifdef MAPBASE_VSCRIPT
 	DEFINE_KEYFIELD( m_iScriptLanguage, FIELD_INTEGER, "vscriptlanguage" ),
-	DEFINE_KEYFIELD( m_iScriptLanguageClient, FIELD_INTEGER, "vscriptlanguage_client" ),
+	//DEFINE_KEYFIELD( m_iScriptLanguageClient, FIELD_INTEGER, "vscriptlanguage_client" ),
 #endif
 	DEFINE_KEYFIELD( m_bColdWorld,		FIELD_BOOLEAN, "coldworld" ),
 
@@ -420,9 +420,6 @@ IMPLEMENT_SERVERCLASS_ST(CWorld, DT_WORLD)
 	SendPropInt		(SENDINFO(m_bColdWorld), 1, SPROP_UNSIGNED ),
 #ifdef MAPBASE
 	SendPropStringT (SENDINFO(m_iszChapterTitle) ),
-#endif
-#ifdef MAPBASE_VSCRIPT
-	SendPropInt		(SENDINFO(m_iScriptLanguageClient), 4 ), // No SPROP_UNSIGNED to allow -1 (disabled)
 #endif
 END_SEND_TABLE()
 
@@ -485,7 +482,7 @@ CWorld::CWorld( )
 
 #ifdef MAPBASE_VSCRIPT
 	m_iScriptLanguage = SL_NONE;
-	m_iScriptLanguageClient = -2;
+	//m_iScriptLanguageClient = -2;
 #endif
 
 	m_bColdWorld = false;
@@ -552,14 +549,6 @@ void CWorld::Spawn( void )
 	Precache( );
 	GlobalEntity_Add( "is_console", STRING(gpGlobals->mapname), ( IsConsole() ) ? GLOBAL_ON : GLOBAL_OFF );
 	GlobalEntity_Add( "is_pc", STRING(gpGlobals->mapname), ( !IsConsole() ) ? GLOBAL_ON : GLOBAL_OFF );
-
-#ifdef MAPBASE_VSCRIPT
-	if (m_iScriptLanguageClient.Get() == -2)
-	{
-		// Clientside language should be regular language by default
-		m_iScriptLanguageClient.Set( m_iScriptLanguage );
-	}
-#endif
 }
 
 static const char *g_DefaultLightstyles[] =
