+++ sp/src/game/shared/vscript_shared.h
@@ -31,6 +31,17 @@ inline bool VScriptRunScript( const char *pszScriptName, bool bWarnMissing = fal
 // Only allow scripts to create entities during map initialization
 bool IsEntityCreationAllowedInScripts( void );
 
+class ISaveRestoreBlockHandler;
+ISaveRestoreBlockHandler *GetVScriptSaveRestoreBlockHandler();
+
+class CBaseEntityScriptInstanceHelper : public IScriptInstanceHelper
+{
+	bool ToString( void *p, char *pBuf, int bufSize );
+	void *BindOnRead( HSCRIPT hInstance, void *pOld, const char *pszId );
+};
+
+extern CBaseEntityScriptInstanceHelper g_BaseEntityScriptInstanceHelper;
+
 #ifdef MAPBASE_VSCRIPT
 void RegisterSharedScriptConstants();
 void RegisterSharedScriptFunctions();
